Move mqtt_json_to_event into models
