+++ tests/test_mqtt.py
@@ -3,17 +3,26 @@
 pytest --cov-report term-missing --cov=axis.mqtt tests/test_mqtt.py
 """
 
+from __future__ import annotations
+
 import json
 from typing import TYPE_CHECKING
 from unittest.mock import MagicMock
 
 import pytest
 
-from axis.interfaces.mqtt import MqttClientHandler, mqtt_json_to_event
-from axis.models.mqtt import ClientConfig, Message, Server, ServerProtocol, Ssl
+from axis.models.mqtt import (
+    ClientConfig,
+    Message,
+    Server,
+    ServerProtocol,
+    Ssl,
+    mqtt_json_to_event,
+)
 
 if TYPE_CHECKING:
     from axis.device import AxisDevice
+    from axis.interfaces.mqtt import MqttClientHandler
 
 
 @pytest.fixture
